// src/shared/enums.ts
export enum RepoResponseStatus {
  SUCCESS = "SUCCESS",
  ERROR = "ERROR",
  NOT_FOUND = "NOT_FOUND",
  ALREADY_EXISTS = "ALREADY_EXISTS"
}

export enum ServiceResponseStatus {
  OK = "OK",
  ERROR = "ERROR",
  ALREADY_EXISTS = "ALREADY_EXISTS",
  NOT_FOUND = "NOT_FOUND",
  INVALID_DATA = "INVALID_DATA",
  BOT_PROMOTED = "BOT_PROMOTED",
  BOT_KICKED = "BOT_KICKED",
  BOT_ADDED = "BOT_ADDED",
  ADD_ADMIN = "ADD_ADMIN",
  REMOVE_ADMIN = "REMOVE_ADMIN"
}